bug fixes and cleanup
