+++ robosurvstratopt/test_specs/oop_demo_test_spec.json
@@ -4,9 +4,10 @@
         "graph_name": "complete_N12",
         "graph_code": "N12_147573952589676412927",
         "num_robots": 1,
-        "objective_function": "ER_pi",
+        "objective_function": "weighted_Stackelberg_pi",
         "eta": 0.25,
         "tau": 10,
+        "B": 100,
         "stationary_distribution": [0.1536, 0.1039, 0.1028, 0.1005, 0.0958, 0.0958, 0.0855, 0.0739, 0.0554, 0.0497, 0.0439, 0.0392],
         "weight_matrix": [[1, 3, 3, 5, 4, 6, 3, 5, 7, 4, 6, 6],
             [3, 1, 5, 4, 2, 4, 4, 5, 5, 3, 5, 5],
@@ -27,7 +28,7 @@
         "num_LCPs": 1,
         "use_abs_param": true,
         "alpha": 1000,
-        "max_iters": 1000,
+        "max_iters": 1,
         "optimizer_name": "rmsprop",
         "nominal_learning_rate": 1,
         "lr_schedule_type": "constant",
